Verify memory alignment rules on loads and stores
Right now all memory accesses are fixed to 4 byte boundries because of the
rules in the register definitions - at some point in the future no doubt that
will change though, and we'll risk storing unaligned words / similar and
getting a poke in the eye with a sharp stick. So, actually check these
conditions.