Scale register-loaded offsets to mem access by appropriate amount
When a value is too large for the offset field in a load/store instruction,
we load it into a register and use that for the offset. However, hardware
still scales that up by the alignment of the memory access. So when creating
instructions to load the offset into a register, first scale it down by the
amount hardware will scale it up by.