Handle all addressing methods no matter what type addr has
LLVM is quite happy letting any expression calculate an address, and that's
sane. However the code here assumed that all memory accesses would be of the
form (somthing+something), or similar arithmatic. Anything vaguely complex
gets ignored and CannotYetSelect'd.

So instead we handle add/subtract addresses as a special case we can optimise,
and allow all inputs to be the calculation of an address.