Rework how we pass predicate information down to SDNodes
This is because LLVM 2.7 panics when it sees us passing around predicates
that actually contain two nodes - an integer and a register. I don't
know the specific change that causes this to occur, but never mind.

It *seems* to be fine to just leave the default predecate operand
generation this way, but accepting them in a pattern and attempting to
stuff them into an instruction node causes tablegen to stomp on an
assertion.

So instead, non-default predicate instructions (IE, those that accept
predicates in a pattern;) Take two individual input operands rather
than the congealed form. This takes the same shape when it reaches
SDNode form.

For whatever reasons a ComplexPattern is needed to disable some typing
rules when juggling the input type in the pattern to a nodefault_reg.