Fixup lowering code for 2.7 changes - largely getLoad/Store changes.
For these, we default to assuming the alignment is 4 bytes, assume that we're
not volatile, and "isNonTemporal" is false. These are all reasonable
assumptions, but there may be special cases to be visited in the future
that I can't forsee right now.

Also add some extra includes, better matching our dependancies