Remunge of asm printer for 2.7:
Switch to using MCAsmInfo names instead of TargetAsmInfo. Use
EmitFunctionHeader to output various fudge and linkage regarding the
start of a function - don't use EmitFunctionBody, as we have to
pump out instructions with predicates manually.

Replace calls to getMangledName with getNameWithPrefix - this seems to
be getMangledNames replacement, but I worry the semantics are different...

Printing .comm/.lcomm situation has changed, with unknown side effects,
so just don't support printing it, we'll run into trouble there when it
arrives.