Seriously rejiggle load/store instruction selection.
So, after removing the code for automagically determining what
side, xpath and datapath an instruction uses, we need to be able to
decide which of these options are required for particular instructions.
On the whole this is pretty simple, because everything uses side A/1
right now.

Unfortunately loads and stores are more complicated - their bases
can be any of the A regs, and on some occasions B15 if we're storing
at the top of the stack (aka, function call setup). So, we now have
to make decisions somewhere in instruction selection inre this.

Most of the time this can be ignored as the register allocator will
only spit out A-registers, but when lowering fixes a physical register
this selection needs to occur. So: we make the relevant side-specific
loads and stores matchable, and put complexpatterns on their address
operands. These then feed into select_addr_regspec, which inspects
and decides whether the selected instruction is appropriate. A clause
in select_addr(_generic) prevents loads/stores with physical register
operands from being autoaccepted into the default forms.

Isn't that increadibly icky? But it's all for the better.