Permit uses of B15 to coalesce for loads/stores, but no other operations
See the large comment; this is because we can cope with loads/stores
relative to B15 through some instruction selection code, but not with
a situation where we end up with an instruction writing its output
to a B-side register. So, all moves to/from B-side that aren't B15
are not recognized as moves, because we can't handle it if they were
coalesced.