Simplify "select" of two values into one instruction
Previously we were making two conditional instructions, one moved the
"true" value in, the other the "false". This only takes one cycle if
you do it in parallel, but we're not. So instead take the simpler route
of one conditional, that overwrites one value with the other if the
condition is met.