Eliminate side and datapath specific load/stores
Hurrah :|. Seeing how we don't write the datapath to the output path,
this means the assembler can work out what the side and datapath
being used are. This isn't dangerous as we're not going for
instruction parallelism right now, so no opportunity for resource
conflicts.

when parallelism does come along, we can re-introduce those side
and datapath specific instructions, but they'll have to be beaten
into code during some later pass, not instruction selection.