Hardcode prolog/epilog as a big string, hack them into machineinstrs.
The reason for this is so that the prolog/epilog instructions can occur
in parallel, two cycles for each plus 4 nops in the epilog for the ret
ptr to load.

The _real_ reason for _why_ we have to do this in parallel is because
of the horror-filled way in which tic64x interrupts interface with the
interrupted context: there's no system stack, so intrs jump on top of
whatever code was running at the time. In our previous prolog/epilog,
there are several cycles where the stack configuration is in a
nonstandard and somewhat undefined state. After an interrupt occurs,
for reasons I can't pin down an exception occurs on or after a load
or store.

I know it's interrupts because disabling them across prolog/epilog
prevents this from happening. I don't know exactly why the presence of
an interrupt during this period causes the fault, but chances are it's
because of the hacky nature of tic64xs interrupt stack and some dodgy
looking code I've disassembled from the short ISR for the mailbox intr
on IVA2.2.

Also, bees.