Implement ReverseBranchCondition, fixes some emission faults LLVM causes.
This is because it appears, if it can't reverse the branch condition,
the BB is still marked as falling through to the next one, even in
cases where we have:
[ A0]  b  .BB1_1
       b  .BB1_2
.BB1_1:
.BB1_2:

This means the BB1_1 label doesn't get written out, and we have an
undefined reference to it. I'm guessing LLVM shouldn't do this if
ReverseBranchCondition isn't implemented/fails, but there we go.